move get prompt to use
